Y'know what? Let the order system rot
I've been coding this and making the already complicated things harder for myself, so I changed the way this works, I ditched the Order and number system, so it's easier for me. I've been working on it for 3 hours and I barely got anywhere, since I finished Acheivement.csv file already, the way I'm doing the Narrator has to be some what wrong. And it's harder to fix later too! Now that I'm made this more understandable to myself, I'm more eager to work